requirements.txt
Flask
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
langchain-google-genai
google-generativeai
chromadb
PyMuPDF
python-docx
numpy
pandas
nltk
sacrebleu
waitress
graphviz
python-pptx
pdfplumber
pymupdf